+++ app.js
@@ -429,6 +429,7 @@ fs.readFile('./config/ipbans.txt', function (err, data) {
 require('./repl.js').start('app', function (cmd) { return eval(cmd); });
 
 //Wulu.Economy.currency_name = 'goat';
+Wulu.UserMod();
 Wulu.Commands.init();
 Wulu.Emoticons();
 Wulu.Tour();