+++ wulu/src/commands/shop.js
@@ -22,53 +22,6 @@ let global_shop = getShopDisplay(shop_data);
  */
 
 function shop(shop=shop_data) {
-  /**
-   * Enabling permanent custom symbols
-   */
-
-  if (!Users.User.prototype.originalJoinRoom) {
-    Users.User.prototype.originalJoinRoom = Users.User.prototype.joinRoom;
-  }
-
-  Users.User.prototype.getIdentity = function(roomid) {
-    if (this.locked) {
-      return '‽' + this.name;
-    }
-    if (roomid) {
-      if (this.mutedRooms[roomid]) {
-        return '!' + this.name;
-      }
-      var room = Rooms.rooms[roomid];
-      if (room && room.auth) {
-        if (room.auth[this.userid]) {
-          return room.auth[this.userid] + this.name;
-        }
-        if (room.isPrivate === true) return ' ' + this.name;
-      }
-    }
-    if (this.customSymbol) {
-      return this.customSymbol + this.name;
-    }
-    return this.group + this.name;
-  };
-
-  Users.User.prototype.joinRoom = function(room, connection) {
-    if (room !== 'global') return this.originalJoinRoom(room, connection);
-    let self = this;
-    // Add delay because when user first join, they don't have there username yet.
-    setTimeout(function() {
-      User.findOne({name: self.userid}, function(err, userModel) {
-        if (err) return;
-        if (userModel && userModel.symbol) {
-          self.customSymbol = userModel.symbol;
-          self.updateIdentity();
-          self.hasPermaCustomSymbol = true;
-        }
-      });
-    }, 1000 * 10);
-    return this.originalJoinRoom(room, connection);
-  };
-
   let commands = {
     shop() {
       if (!this.canBroadcast()) return;
@@ -144,6 +97,7 @@ function shop(shop=shop_data) {
       this.sendReply('Your symbol has been reset.');
     },
 
+    permasymbol: 'permacustomsymbol',
     permacustomsymbol(target, room, user) {
       if (!user.canPermaCustomSymbol) return this.sendReply('You need to buy this item from the shop.');
       if (target.length > 1) return this.sendReply('/permacustomsymbol [symbol] - Get a custom symbol.');